requirements.txt
aiosmtplib==3.0.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
babel==2.17.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
captcha==0.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
Django==5.1.6
django-ranged-response==0.2.0
djangorestframework==3.15.2
dnspython==2.7.0
docutils==0.21.2
dotenv==0.9.9
dulwich==0.22.8
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-limiter==0.1.6
fastapi-mail==1.4.2
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imagesize==1.4.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
libgravatar==1.0.4
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
mysqlclient==2.2.7
numpy==2.2.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pbs-installer==2025.2.12
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.1
pywin32-ctypes==0.2.3
RapidFuzz==3.12.2
redis==5.3.0b5
requests==2.32.3
requests-toolbelt==1.0.0
roman-numerals-py==3.1.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.46.0
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.2.18.16
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
zstandard==0.23.0
